// src/service/main/system/sysButton/model.d.ts
export interface SysButtonDto {
  id: string;
  menuId: string;
  code: string;
  name: string;
  icon: string;
  showMode: string;
  btnType: string;
  state: boolean;
  createdBy: string;
  createdAt: string;
  updatedBy: string;
  updatedAt: string;
  deletedBy: string;
  deletedAt: string;
}

export interface SysButtonVo {
  id: string;
  menuId: string;
  code: string;
  name: string;
  icon: string;
  showMode: string;
  btnType: string;
  state: boolean;
  createdBy: string;
  createdAt: string;
  updatedBy: string;
  updatedAt: string;
  deletedBy: string;
  deletedAt: string;
}

export interface SysButtonQuery {
  id: string;
  menuId: string;
  code: string;
  name: string;
  icon: string;
  showMode: string;
  btnType: string;
  state: boolean;
  createdBy: string;
  createdAt: string;
  updatedBy: string;
  updatedAt: string;
  deletedBy: string;
  deletedAt: string;
}